Fix build check script
Fix an issue where grep exited with exit code 1 when no lines matched
the final pattern.  Not matching any lines is the expected behavior, but
the variable should be populated with empty string instead of causing
the shell to exit (because of `set -e`).

In addition, do not install dev dependencies since they are not needed