Simplify build process by removing built site from version control
Previously, the site was served via the `docs` folder checked into
version control.  This required building the site and checking it in
every time anything changed.

This commit changes this process by removing the built site from version
control in favor of building and deploying it on each merge to main.

It is still possible to build and preview the site locally.  The only
difference is the site is built into and served from the "dist" folder
instead of the "docs" folder, using a much more predictable folder name.